scraper/requirements.txt
requests
beautifulsoup4
pandas
kafka-python
